Debug de l'implementation du raytracer

See merge request anthony_rolland/sae1.01-groupe13!22

+++ app/src/main/java/sae101/parser/scene/Builder.java
@@ -20,7 +20,7 @@ public interface Builder {
      * @param height the height
      * @param width  the width
      */
-    void setDimensions(float height, float width);
+    void setDimensions(int height, int width);
 
     /**
      * Add light.
